+++ Changelog.md
@@ -1,6 +1,7 @@
-# 0.2.7
+# Next
 
 - Update `updateLastConsumedMessageIndex` on the channel when a user has viewed a message.
+- Improved error handling (If I'm lucky).
 
 
 # 0.2.6